Connecting to Drive
New function made that creates a drive builder object, the object that connects to the google drive. Now the connection is made inside the FXML controller, so only one connection to google drive is needed.

Not ideal, but the functions need the drive builder object to be made in the same file as where they are called, and as far as I can tell, they need to be called inside the FXML controller for buttons to work, so it'll do, its not that bad.